rename python path
