+++ .azure-pipelines/scripts/install_nc.sh
@@ -5,7 +5,7 @@ cd /neural-compressor
 
 python -m pip install --no-cache-dir -r requirements.txt
 python setup.py bdist_wheel
-pip install dist/neural_compressor*.whl --force-reinstall
+pip install dist/onnx_neural_compressor*.whl --force-reinstall
 
 echo -e "\n pip list after install Neural Compressor ... "
 pip list
