+++ docs/quantization_layer_wise.md
@@ -49,7 +49,7 @@ Large language models (LLMs) have shown exceptional performance across various t
 ## Examples
 
 ```python
-from neural_compressor_ort.quantization import matmul_4bits_quantizer
+from onnx_neural_compressor.quantization import matmul_4bits_quantizer
 
 algo_config = matmul_4bits_quantizer.RTNWeightOnlyQuantConfig(layer_wise_quant=True)
 quant = matmul_4bits_quantizer.MatMul4BitsQuantizer(
