+++ onnx_neural_compressor/algorithms/__init__.py
@@ -12,6 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from neural_compressor_ort.algorithms.smoother.core import Smoother
 
-__all__ = ["Smoother"]
+from onnx_neural_compressor.algorithms.smoother import Smoother
+from onnx_neural_compressor.algorithms.weight_only.rtn import apply_rtn_on_model
+from onnx_neural_compressor.algorithms.weight_only.gptq import apply_gptq_on_model
+from onnx_neural_compressor.algorithms.weight_only.awq import apply_awq_on_model
+from onnx_neural_compressor.algorithms.layer_wise import layer_wise_quant
+
+__all__ = ["Smoother", "apply_rtn_on_model", "apply_gptq_on_model", "apply_awq_on_model", "layer_wise_quant"]
