+++ onnx_neural_compressor/algorithms/weight_only/awq.py
@@ -26,12 +26,12 @@ import onnx
 import onnxruntime as ort
 from packaging.version import Version
 
-from neural_compressor_ort.algorithms.weight_only.rtn import rtn_quantize
-from neural_compressor_ort.algorithms.weight_only.utility import pad_tensor, prepare_inputs, qdq_tensor
-from neural_compressor_ort.quantization.calibrate import CalibrationDataReader
-from neural_compressor_ort.quantization.config import AWQConfig
-from neural_compressor_ort.utils import ONNXRT116_VERSION, ONNXRT1161_VERSION, dtype_mapping, logger
-from neural_compressor_ort.utils.onnx_model import ONNXModel
+from onnx_neural_compressor.algorithms.weight_only.rtn import rtn_quantize
+from onnx_neural_compressor.algorithms.weight_only.utility import pad_tensor, prepare_inputs, qdq_tensor
+from onnx_neural_compressor.quantization.calibrate import CalibrationDataReader
+from onnx_neural_compressor.quantization.config import AWQConfig
+from onnx_neural_compressor.utils import ONNXRT116_VERSION, ONNXRT1161_VERSION, dtype_mapping, logger
+from onnx_neural_compressor.utils.onnx_model import ONNXModel
 
 __all__ = ["apply_awq_on_model", "awq_quantize"]
