+++ onnx_neural_compressor/algorithms/weight_only/gptq.py
@@ -26,16 +26,16 @@ import onnx
 import onnxruntime as ort
 from packaging.version import Version
 
-from neural_compressor_ort.algorithms.weight_only.utility import (
+from onnx_neural_compressor.algorithms.weight_only.utility import (
     make_matmul_weight_only_node,
     pad_tensor,
     prepare_inputs,
     quant_tensor,
 )
-from neural_compressor_ort.quantization.calibrate import CalibrationDataReader
-from neural_compressor_ort.quantization.config import GPTQConfig
-from neural_compressor_ort.utils import ONNXRT116_VERSION, ONNXRT1161_VERSION, dtype_mapping, simple_progress_bar
-from neural_compressor_ort.utils.onnx_model import ONNXModel
+from onnx_neural_compressor.quantization.calibrate import CalibrationDataReader
+from onnx_neural_compressor.quantization.config import GPTQConfig
+from onnx_neural_compressor.utils import ONNXRT116_VERSION, ONNXRT1161_VERSION, dtype_mapping, simple_progress_bar
+from onnx_neural_compressor.utils.onnx_model import ONNXModel
 
 __all__ = [
     "apply_gptq_on_model",
@@ -426,7 +426,7 @@ def apply_gptq_on_model(
         if isinstance(op_config, GPTQConfig):
             quant_config[op_name_type] = op_config.to_dict()
     if layer_wise:
-        from neural_compressor_ort.algorithms import layer_wise_quant
+        from onnx_neural_compressor.algorithms import layer_wise_quant
 
         quantized_model = layer_wise_quant(
             model,
