+++ test/quantization/test_config.py
@@ -7,7 +7,7 @@ import numpy as np
 import onnx
 from optimum.exporters.onnx import main_export
 
-from neural_compressor_ort.utils import logger
+from onnx_neural_compressor.utils import logger
 
 
 def find_onnx_file(folder_path):
@@ -82,8 +82,8 @@ class TestQuantizationConfig(unittest.TestCase):
         return len(op_names)
 
     def test_config_white_lst(self):
-        from neural_compressor_ort.quantization import RTNConfig
-        from neural_compressor_ort.quantization.quantize import _quantize
+        from onnx_neural_compressor.quantization import RTNConfig
+        from onnx_neural_compressor.quantization.quantize import _quantize
 
         global_config = RTNConfig(weight_bits=4)
         # set operator instance
@@ -96,8 +96,8 @@ class TestQuantizationConfig(unittest.TestCase):
         self.assertFalse(self._check_node_is_quantized(qmodel, "/h.4/mlp/fc_out/MatMul"))
 
     def test_config_white_lst2(self):
-        from neural_compressor_ort.quantization import RTNConfig
-        from neural_compressor_ort.quantization.quantize import _quantize
+        from onnx_neural_compressor.quantization import RTNConfig
+        from onnx_neural_compressor.quantization.quantize import _quantize
 
         global_config = RTNConfig(weight_dtype="fp32")
         # set operator instance
@@ -110,8 +110,8 @@ class TestQuantizationConfig(unittest.TestCase):
         self.assertTrue(self._check_node_is_quantized(qmodel, "/h.4/mlp/fc_out/MatMul"))
 
     def test_config_white_lst3(self):
-        from neural_compressor_ort.quantization import RTNConfig
-        from neural_compressor_ort.utils.utility import get_model_info
+        from onnx_neural_compressor.quantization import RTNConfig
+        from onnx_neural_compressor.utils.utility import get_model_info
 
         global_config = RTNConfig(weight_bits=4)
         # set operator instance
@@ -127,7 +127,7 @@ class TestQuantizationConfig(unittest.TestCase):
         self.assertTrue(configs_mapping[("/h.4/mlp/fc_in/MatMul", "MatMul")].weight_bits == 4)
 
     def test_config_from_dict(self):
-        from neural_compressor_ort.quantization import RTNConfig
+        from onnx_neural_compressor.quantization import RTNConfig
 
         quant_config = {
             "rtn": {
@@ -148,7 +148,7 @@ class TestQuantizationConfig(unittest.TestCase):
         self.assertIsNotNone(config.local_config)
 
     def test_config_to_dict(self):
-        from neural_compressor_ort.quantization import RTNConfig
+        from onnx_neural_compressor.quantization import RTNConfig
 
         quant_config = RTNConfig(weight_bits=4)
         fc_out_config = RTNConfig(weight_bits=8)
@@ -158,7 +158,7 @@ class TestQuantizationConfig(unittest.TestCase):
         self.assertIn("local", config_dict)
 
     def test_same_type_configs_addition(self):
-        from neural_compressor_ort.quantization import RTNConfig
+        from onnx_neural_compressor.quantization import RTNConfig
 
         quant_config1 = {
             "rtn": {
@@ -191,8 +191,8 @@ class TestQuantizationConfig(unittest.TestCase):
         self.assertNotEqual(q3_dict["global"]["weight_bits"], quant_config2["rtn"]["global"]["weight_bits"])
 
     def test_config_mapping(self):
-        from neural_compressor_ort.quantization import RTNConfig
-        from neural_compressor_ort.utils.utility import get_model_info
+        from onnx_neural_compressor.quantization import RTNConfig
+        from onnx_neural_compressor.utils.utility import get_model_info
 
         quant_config = RTNConfig(weight_bits=4)
         # set operator instance
@@ -217,7 +217,7 @@ class TestQuantizationConfig(unittest.TestCase):
         self.assertTrue(configs_mapping[("/h.1/mlp/fc_out/MatMul", "MatMul")].weight_bits == 3)
 
     def test_diff_types_configs_addition(self):
-        from neural_compressor_ort.quantization import GPTQConfig, RTNConfig
+        from onnx_neural_compressor.quantization import GPTQConfig, RTNConfig
 
         quant_config1 = {
             "rtn": {
@@ -237,7 +237,7 @@ class TestQuantizationConfig(unittest.TestCase):
 class TestQuantConfigForAutotune(unittest.TestCase):
     def test_expand_config(self):
         # test the expand functionalities, the user is not aware it
-        from neural_compressor_ort.quantization import RTNConfig
+        from onnx_neural_compressor.quantization import RTNConfig
 
         tune_config = RTNConfig(weight_bits=[4, 8])
         expand_config_list = RTNConfig.expand(tune_config)
