+++ test/utils/test_general.py
@@ -3,17 +3,17 @@
 import unittest
 from typing import Any, List, Optional, Tuple, Union
 
-from neural_compressor_ort.utils import logger
-from neural_compressor_ort.utils.base_config import (
+from onnx_neural_compressor.utils import logger
+from onnx_neural_compressor.utils.base_config import (
     BaseConfig,
     config_registry,
     get_all_config_set_from_config_registry,
     register_config,
     register_supported_configs,
 )
-from neural_compressor_ort.utils.base_tuning import ConfigLoader, ConfigSet, Evaluator, SequentialSampler
-from neural_compressor_ort.utils.constants import DEFAULT_WHITE_LIST, OP_NAME_OR_MODULE_TYPE
-from neural_compressor_ort.utils.tuning_param import TuningParam
+from onnx_neural_compressor.utils.base_tuning import ConfigLoader, ConfigSet, Evaluator, SequentialSampler
+from onnx_neural_compressor.utils.constants import DEFAULT_WHITE_LIST, OP_NAME_OR_MODULE_TYPE
+from onnx_neural_compressor.utils.tuning_param import TuningParam
 
 PRIORITY_FAKE_ALGO = 100
 FAKE_CONFIG_NAME = "fake"
