+++ test/utils/test_logger.py
@@ -20,7 +20,7 @@ log_msg_lst = [
 
 class TestLogger(unittest.TestCase):
     def test_logger(self):
-        from neural_compressor_ort.utils import logger
+        from onnx_neural_compressor.utils import logger
 
         for msg in log_msg_lst:
             logger.log(0, msg)
@@ -52,7 +52,7 @@ class TestLogger(unittest.TestCase):
         # logger.warning(({"msg": [{"sub_msg": "call logger"}, {"sub_msg2": "call warning function"}]}, {"msg2": "done"}))
 
     # def test_logger_func_and_pretty_dict(self):
-    #     from neural_compressor_ort.utils import debug, error, fatal, info, log, warning
+    #     from onnx_neural_compressor.utils import debug, error, fatal, info, log, warning
 
     #     for msg in log_msg_lst:
     #         log(0, msg)
